v 1.05
Rewrote the notebook so that equation numbering was not used in the text; this avoids the problem caused by different rendering engines doing different things (sometimes numbering equations, and sometimes not).  Here I did put in one equation with the \begin{equation*} \end{equation*} environment and indeed sometimes this is not numbered when other equations are.